Fixed issues with list bullets on page break
There are several issues mentioning problems with the rendering of bullets when the page is breaked within a list (see #403, #872 and #1045).

When dompdf detects that the content of a list item overflows the page, it clones the item to the next page and moves all following content to it. If the first child of the item leads to the overflow, the bullet is still drawn at the end of the first page and no bullet is drawn for the cloned item on the next page.

In fact there are three issues that I address in this PR:
* The origin list item gets removed if it has no content anymore after it was splitted.
* The cloned element is marked as non-splitted if the origin item was removed, so that the bullet of the cloned item on the next page gets rendered.
* A new bullet element is created on the cloned list item, but the attribute `dompdf-counter` is not filled correctly, so that the bullet is not rendered if an ordered list (`<ol>`) is used.

The fix is a bit hacky, but I couldn't find a cleaner way of fixing the issue. The problem is that the bullet is handled by a custom DOM element which makes some checks more complicated.

Maybe this could be changed sometime.